Revert "refactor: reorg files in backend (#2449)"
This reverts commit dc6bd5351945e6a256e26382b8f2a433e72520d2.
