+++ backend/llm/utils/call_brain_api.py
@@ -1,10 +1,10 @@
 from uuid import UUID
 
 from fastapi import HTTPException
-from utils.extract_api_brain_definition_values_from_llm_output import (
+from llm.utils.extract_api_brain_definition_values_from_llm_output import (
     extract_api_brain_definition_values_from_llm_output,
 )
-from utils.make_api_request import get_api_call_response_as_text
+from llm.utils.make_api_request import get_api_call_response_as_text
 from modules.brain.service.api_brain_definition_service import ApiBrainDefinitionService
 from modules.brain.service.brain_service import BrainService
