+++ backend/llm/utils/get_api_brain_definition_as_json_schema.py
@@ -1,6 +1,6 @@
 from fastapi import HTTPException
-from utils.extract_api_definition import format_api_brain_property
-from utils.sanitize_function_name import sanitize_function_name
+from llm.utils.extract_api_definition import format_api_brain_property
+from llm.utils.sanitize_function_name import sanitize_function_name
 from modules.brain.entity.brain_entity import BrainEntity
 from modules.brain.service.api_brain_definition_service import ApiBrainDefinitionService
