+++ backend/models/databases/repository.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from datetime import datetime
 from uuid import UUID
 
-from .llm_models import LLMModels
+from .entity import LLMModels
 
 
 class Repository(ABC):
