+++ backend/modules/brain/api_brain_qa.py
@@ -6,8 +6,8 @@ import jq
 import requests
 from fastapi import HTTPException
 from litellm import completion
-from utils.call_brain_api import call_brain_api
-from utils.get_api_brain_definition_as_json_schema import (
+from llm.utils.call_brain_api import call_brain_api
+from llm.utils.get_api_brain_definition_as_json_schema import (
     get_api_brain_definition_as_json_schema,
 )
 from logger import get_logger
