+++ backend/modules/brain/knowledge_brain_qa.py
@@ -3,12 +3,12 @@ from typing import AsyncIterable, List, Optional
 from uuid import UUID
 
 from langchain.callbacks.streaming_aiter import AsyncIteratorCallbackHandler
-from utils.format_chat_history import format_chat_history
-from utils.get_prompt_to_use import get_prompt_to_use
-from utils.get_prompt_to_use_id import get_prompt_to_use_id
+from llm.utils.format_chat_history import format_chat_history
+from llm.utils.get_prompt_to_use import get_prompt_to_use
+from llm.utils.get_prompt_to_use_id import get_prompt_to_use_id
 from logger import get_logger
 from models import BrainSettings
-from modules.user.service.user_usage import UserUsage
+from models.user_usage import UserUsage
 from modules.brain.entity.brain_entity import BrainEntity
 from modules.brain.qa_interface import QAInterface
 from modules.brain.rags.quivr_rag import QuivrRAG
@@ -24,7 +24,7 @@ from modules.chat.dto.outputs import GetChatHistoryOutput
 from modules.chat.service.chat_service import ChatService
 from pydantic import BaseModel, ConfigDict
 from pydantic_settings import BaseSettings
-from modules.upload.service.generate_file_signed_url import generate_file_signed_url
+from repository.files.generate_file_signed_url import generate_file_signed_url
 
 logger = get_logger(__name__)
 QUIVR_DEFAULT_PROMPT = "Your name is Quivr. You're a helpful assistant.  If you don't know the answer, just say that you don't know, don't try to make up an answer."
