+++ backend/repository/brain_subscription/__init__.py
@@ -1,2 +1,3 @@
+from .get_brain_url import get_brain_url
 from .resend_invitation_email import resend_invitation_email
 from .subscription_invitation_service import SubscriptionInvitationService
